// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const PullRequestStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  SUMMARIZED: 'SUMMARIZED'
} as const

export type PullRequestStatus = (typeof PullRequestStatus)[keyof typeof PullRequestStatus]


export const SubscriptionStatus = {
  pending: 'pending',
  authenticated: 'authenticated',
  active: 'active',
  completed: 'completed',
  paused: 'paused',
  cancelled: 'cancelled',
  halted: 'halted'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const PaymentStatus = {
  pending: 'pending',
  captured: 'captured',
  failed: 'failed'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const InvitationStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected'
} as const

export type InvitationStatus = (typeof InvitationStatus)[keyof typeof InvitationStatus]